Fixed RoleInstance log value on Search Service (#194)
The search service used to use ServerCommon's application insights helper to set the application key, which happened to also overwrite the cloud role instance value to the local machine's name. This is not the intended behavior on Azure. This change allows Application Insights to properly initialize the default values, and then sets the local machine's name as a custom dimension on Application Insight telemetry items.