Introduce IndexOperationBuilder to encapsulate query optimizations (#604)
This introduces a new class which will encapsulate this sort of logic.
It will parse the query, inspect the groupings, and generate the parameters for the appropriate operation.
For "lookup document" it will generate the document key and for "search" it will generate the search text and search parameters.
Progress on https://github.com/NuGet/NuGetGallery/issues/6466