[Azure Search] Add option to skip version list initialization (#753)
A good chunk of time is spent by Db2AzureSearch to initialize the internal Version List resource. This resource is used by catalog2azuresearch to update the Azure Search index, thus, it is unnecessary if an advanced user just wants to create a search index without keeping it up-to-date. This change adds an config to disable version list initialization to speed up db2azuresearch. On DEV, this reduces the total execution time by ~8.4%. 

This change also moves the existing `ReplaceContainersAndIndexes` and `SkipPackagePrefixes` into a new `Development` configuration section.