[Package Renames 1] Add popularity transfers data client (#765)
The `auxiliary2azuresearch` job needs to know which popularity transfers have changed to properly update the search index. To do this, the last seen popularity transfers data need to be persisted so that it can be compared against the latest data.

This adds new types `IPopularityTransferDataClient` and `PopularityTransferDataClient` that are unused for now. They will be added to the dependency injection container later.

Part of https://github.com/nuget/nugetgallery/issues/7898